scraper/requirements.txt
pymongo==3.11
websocket-client==1.1.0
requests==2.25.1
